+++ src/shared/ui/Navbar/Navbar.tsx
@@ -3,7 +3,7 @@ import logo from "../../assets/images/logopurple.webp";
 import { BecomeMentorButton } from "../../../features/Mentor/BecomeMentorButton";
 
 export default function Navbar() {
-  const isAuthorized = false;
+  const isAuthorized = true;
   function testFunction() {
     console.log("user click join us");
   }
